+++ CHANGELOG.md
@@ -4,3 +4,5 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+* **201-06-28**: Added Admin for routing bundle
+* **201-06-09**: Added Admin for seo-bundle